requirements.txt
playwright
beautifulsoup4
lxml
requests
python-dotenv
html5lib
